fix for automatic selection of tier

+++ database/rrd.h
@@ -401,7 +401,7 @@ struct rrddim_tier {
     size_t iterations;
     time_t next_point_time;
     size_t anomaly_count;
-    uint8_t tier_grouping;
+    int tier_grouping;
     usec_t last_collected_ut;
     struct rrddim_collect_ops collect_ops;
     struct rrddim_query_ops query_ops;
